requirements.txt
mysql-connector-python==8.0.34
